# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OmniAudit API
  description: |
    Enterprise-Grade Universal AI Coding Optimization Framework API.

    OmniAudit provides comprehensive code analysis capabilities including:
    - Security vulnerability detection
    - Code quality analysis
    - Performance optimization suggestions
    - Dependency vulnerability scanning
    - AI-powered insights and recommendations

    ## Authentication
    All API endpoints require authentication using Bearer tokens.
    Include your API key in the Authorization header:
    ```
    Authorization: Bearer sk-your-api-key
    ```

    ## Rate Limiting
    - Standard tier: 1000 requests/hour
    - Professional tier: 5000 requests/hour
    - Enterprise tier: Unlimited

    ## Webhooks
    Subscribe to events to receive real-time notifications about audit completions,
    critical findings, and more.

  version: 2.0.0
  contact:
    name: OmniAudit Support
    url: https://omniaudit.dev/support
    email: support@omniaudit.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.omniaudit.dev/v1
    description: Production API
  - url: https://staging-api.omniaudit.dev/v1
    description: Staging API
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /audit:
    post:
      summary: Create new audit
      description: Submit a new code audit request
      operationId: createAudit
      tags:
        - Audits
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
      responses:
        '201':
          description: Audit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /audit/{auditId}:
    get:
      summary: Get audit status
      description: Retrieve the status and results of an audit
      operationId: getAudit
      tags:
        - Audits
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AuditId'
      responses:
        '200':
          description: Audit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /batch/audit:
    post:
      summary: Create batch audit
      description: Submit multiple repositories for batch analysis
      operationId: createBatchAudit
      tags:
        - Audits
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchAuditRequest'
      responses:
        '202':
          description: Batch audit queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAuditResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key authentication
  parameters:
    AuditId:
      name: auditId
      in: path
      required: true
      schema:
        type: string
      description: Unique audit identifier
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    AuditRequest:
      type: object
      required:
        - project_path
      properties:
        project_path:
          type: string
          description: Path to the project to audit
        analyzers:
          type: array
          items:
            type: string
            enum: [security, quality, performance, dependencies, accessibility]
          description: Analyzers to run
        config:
          type: object
          properties:
            severity_filter:
              type: array
              items:
                type: string
            max_findings:
              type: integer
            include_suggestions:
              type: boolean
    AuditResponse:
      type: object
      properties:
        audit_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        created_at:
          type: string
          format: date-time
    AuditDetails:
      allOf:
        - $ref: '#/components/schemas/AuditResponse'
        - type: object
          properties:
            completed_at:
              type: string
              format: date-time
            duration_ms:
              type: integer
            findings_count:
              type: integer
            findings_by_severity:
              type: object
              properties:
                critical:
                  type: integer
                high:
                  type: integer
                medium:
                  type: integer
                low:
                  type: integer
                info:
                  type: integer
...
    BatchAuditRequest:
      type: object
      required:
        - repositories
      properties:
        repositories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              path:
                type: string
              url:
                type: string
        analyzers:
          type: array
          items:
            type: string
    BatchAuditResponse:
      type: object
      properties:
        job_id:
          type: string
        total_repos:
          type: integer
        status:
          type: string
          enum: [queued, running, completed]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object